requirements.txt
# Function dependencies, for example:
# package>=version
numpy
requests
Pillow
twilio